+++ packages/native/src/Space/index.ts
@@ -1,1 +1,2 @@
 export * from './Space';
+export * from './types';