+++ website/src/config/index.ts
@@ -13,6 +13,7 @@ export const PATREON_LINK = 'https://www.patreon.com/beautydesign';
 export const PAYPAL_LINK = 'https://www.paypal.com/paypalme/rivaslive';
 export const COINBASE_WALLET = '33RVxLJzVCrzuiNrJY4J4nSwVNVusw4pm7';
 export const LIB_VERSION = pkg.version;
+export const REPOSITORY_URL = pkg?.repository?.url?.replace('git+', '') || '';
 export const libTheme = {
   colorsDark,
   colorsLight,